==> Notebooks/00-instructions.ipynb <==
!pip list --format=freeze > ../requirements.txt

# 0. Preparación

## 0.1 Instalación de módulos

Se instalan los módulos específicados en el archivo requirement.txt para la correcta ejecución del códugo a utilizar.

!pip install -r ../requirements.txt

## 0.2 Obtención del dataset

El dataset puede ser descargado del siguiente link:

https://www.kaggle.com/datasets/prevek18/ames-housing-dataset

## 0.3 Demostraciones

Sea un modelo de regresión lineal simple

$$
y_i  = \beta_0 + \beta_1 x_i+ \epsilon_i,  \forall i = 1,...,n
$$

donde los errores aleatorios cumplen $E(\epsilon_i) = 0$ y $VAR(\epsilon_i) = \sigma^2$,

Demuestra que la suma de cuadrados de los residuos dividida por $\sigma^2$,

$$
\frac{SS_{Res}}{\sigma^2} = \frac{\sum e_i^2}{\sigma^2}
$$

puede expresarse como una combinación cuadrática de los errores $\epsilon_i$

### Paso a Paso

$$
y = X\beta + \epsilon
$$

De la predicción OLS, tenemos que:

$$
\hat{y} = X \beta = X(X'X)^{-1}X'y = Hy
$$

donde $H = X(X'X)^{-1}X'$

Entonces, tenemos que:
$$
r = y - \hat{y} = (I - H)y
$$

reemplazando y con $y = X\beta + \epsilon$:

$$
r = (I - H)(X\beta + \epsilon) = (I - H)X\beta + (I - H)\epsilon
$$

---
Se puede demostrar que $(I-H)X = 0$, ya que:

$$
(I-H)X = IX - HX = X - X = 0
$$

ya que:

$$
HX = X(X'X)^{-1}X'X = X ((X'X)^{-1}(X'X)) = XI = X
$$

---
Entonces:

$$
r = (I - H)X\beta + (I - H)\epsilon = (I - H)\epsilon
$$

También tenemos que:

$$
\begin{aligned}
SSE
&= r'r \\
&= \epsilon'(I - H)'(I - H)\epsilon
\end{aligned}
$$

---
Se puede demostrar que $(I - H)'(I - H) = (I - H)$:

$$
\begin{aligned}
H^2
&= X(X'X)^{-1}X'X(X'X)^{-1}X' \\
&= X((X'X)^{-1}(X'X))(X'X)^{-1}X' \\
&= XI_p(X'X)^{-1}X' \\
&= X(X'X)^{-1}X' \\
&= H \\
\end{aligned}
$$

ya que $((X'X)^{-1}(X'X)) = I$

Tambien tenemos que:
$$
H' = (X(X'X)^{-1}X')' = X(X'X)^{-1}X' = H
$$

Por tanto,
$$
\begin{aligned}
(I - H)'(I - H)
&= (I' - H')(I - H) \\
&= (I - H)(I - H) \\
&= I^2 - 2H + H^2 \\
&= I - 2H + H \\
&= I - H \\
\end{aligned}
$$

---
En base a lo anterior,

$$
\begin{aligned}
SSE
&= \epsilon'(I - H)'(I - H)\epsilon \\
&= \epsilon'(I - H)\epsilon
\end{aligned}
$$

Una combinación cuadratica de las componentes $\epsilon_i$.

Dividiendo entre $\sigma^2$:

$$
\frac{SSE}{\sigma^2} = \frac{1}{\sigma^2} \epsilon'(I - H)\epsilon
$$

---
Entonces, si asumimos que $\epsilon$ es normal, con $\mu = 0$ y $Var = \sigma^2$, y ademàs, dado que H es simétrica e idempotente (como ya ha sido demostrado) tenemos que existe $Q$ tal que:

$$
H = Q'\Lambda Q
$$

Donde $\Lambda = diag(1,1,1,...(n-p) veces,1,1,0,0,...p veces...,0,0)$

por tanto:

$$
\frac{1}{\sigma^2}\epsilon'(I - H)\epsilon = \frac{1}{\sigma^2}\epsilon'Q\Lambda Q'\epsilon = ( \frac{1}{\sigma}Q'\epsilon)'\Lambda(\frac{1}{\sigma}Q'\epsilon) = z'\Lambda z
$$

Dado que Q es ortogonal y $\epsilon$ es normal, $z \sim N$. Entonces:

$$
z'\Lambda z = \sum^{n-p}_{i=1} z_i^2 \sim \chi^2_{n-p}
$$

Por tanto,

$$
\frac{SS_{Res}}{\sigma^2} \sim \chi^2_{n-p}
$$
