==> Notebooks/05-inference.ipynb <==
# 5. Inferencia y grados de libertad

Importamos los datos y módulos a usar.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle  

import statsmodels.api as sm
from statsmodels.regression.linear_model import OLS

import warnings
warnings.filterwarnings('ignore')

path = '../Data/'

X = np.load(path + 'X.dat', allow_pickle= True)
B = np.load(path + 'B.dat', allow_pickle= True)
Y = np.load(path + 'Y.dat', allow_pickle= True)

result =  sm.load(path + 'ols_results.pickle')

with open("../Data/regressor_columns.pickle", "rb") as fp:   #Pickling
...   regressor_columns = pickle.load(fp)

df = pd.read_csv(path+'Ames-Housing-regressor-columns.csv', delimiter=',')

## 5.1 Calculo de grados de libertad

Los grados de libertad de modelo corresponden a número de coeficientes B estimados, asociados a variables regresoras. Es decir, es calculado como:

$$Df_{Model} = m - 1 = k$$

Df_Model = len(B) - 1

Df_Model

result.df_model

Por su parte, los grados de libertad de los residuos corresponden al número de observaciones n menos el total de coeficientes estimados:

$$Df_{Residuals} = n - m = n - (k + 1)$$

Df_Residuals = len(Y) - len(B)

Df_Residuals

result.df_resid

Comparando con los grados de libertad total, que se calcula como n - 1, por la perdida de un grado de libertad por la estimación de $\overline{Y}$:

$$Df_{Total} = n - 1$$

Por lo que se cumple que:

$$Df_{Total} = n - 1 - k + k$$
$$Df_{Total} = n - (k + 1) + k$$
$$Df_{Total} = Df_{Residuals} + Df_{Model}$$

Df_Total = len(Y) - 1

Df_Total

Df_Residuals + Df_Model

## 5.2 Calculo de error estandar

Para calcular el error estándar de cada coeficiente, primero calculamos el error medio cuadrático, usando la expresión:

$$MSE = \sum \frac{(\hat{y} - y)^2}{df_{residuals}}$$

Y_pred = X @ B

MSR = np.sum((Y_pred - Y)**2)/(Df_Residuals)

MSR

Luego, se obtiene la matriz de covarianza de los estimadores calculada como:

$$COV = (X^T X)$$

De esta matriz, nos interesa los valores que están en la diagonal central, los cuales corresponderán a la covarianza de los estimadores, incluyendo $B_0$.

Además, dado que para calcular el error estandar es necesario dividir el error cuadratico medio sobre los valores de la covarianza, es necesario sacar su inversa. De esta manera calculamos el error estandar como:

cov_beta = MSR * np.linalg.inv(X.T @ X)

std_error = np.sqrt(np.diag(cov_beta))

std_error
    

## 5.3 Testeo de hipótesis y p-valor

Con el error estándar podemos obtener un estadístico t, que nos ayude a validar la hipotesis de significancia del modelo, donde:

$$H_0: B_i = 0, \forall i > 0$</p> <p style="text-align:center;">$H_1: B_i \ne 0, \forall i > 0$$

Se entiende que si no se puede rechazar la hipotesis nula que el coeficiente asociado a un atributo $X_i$ es igual a 0, entonces ese atributo no tiene una significancia verdadera en el modelo, por lo cual puede ser descartado. El estadistico t asociado será igual a:
$$t_i = \frac{B_i}{se_i}, \forall i > 0$$

Y el p valor será igual a:
$$p_i = 2 * P(t>|t_i|), \forall i > 0$$

Una hipótesis similar se puede realizar para el intercepto $B_0$, donde se evalúa la hipótesis de que el modelo pase por el origen, es decir, $Y=0$ para $X = 0$.

## 5.4 Tabla de coeficientes

Obtenemos la tabla de coeficientes otorgada por statmodels. Comparamos los errores estándares con los calculados y verificamos que efectivamente obtuvimos los mismos resultados:

result.summary()

Analizando la tabla, llegamos a las siguientes conclusiones:
* El intercepto obtenido es negativo, con valor estimado de -1.2768, y se rechaza la hipótesis nula que sea 0.
* Los parámetros $B_1, B_2, B_4, B_5, B_6, B_8$ y $B_{10}$ son positivos, con p-valor inferior a 5%, por lo que se rechaza la hipótesis nula que sean 0 y no tengan significancia en el modelo. Que sean negativos implica un incremento en el atributo $X_i$ implica un aumento en el precio de venta.
  
* Los parámetros $B_7, B_9, B_{11}, B_{12}, B_{13}$ y $B_{14}$ son negativos, con p-valor inferior a 5%, por lo que se rechaza la hipótesis nula que sean 0 y no tengan significancia en el modelo. Que sean negativos implica un incremento en el atributo $X_i$ implica una reducción en el precio de venta.
  
* Los parámetros $B_3, B_{10}$ tienen un p-valor superior al 5%, por lo que no se puede rechazar la hipótesis nula de que es igual a 0. Por tanto, podemos asumir que su significancia en el modelo es nula, en comparación a los otros parámetros.

* Algunos coeficientes, como $B_4$, presentan un valor cercano a 0, aunque el p-valor sigue siendo inferior al 5%. Esto puede deberse a que el atributo $X_4$ se mueve en un rango de valores altos, lo que hace que su coeficiente asociado sea bajo, pero sin perder su significancia.

* El p-valor para el intercepto no nos permite rechazar la hipótesis nula de que sea distinto de 0, por ende, se puede asumir que el origen hace parte de la superficie descrita por el modelo.

## 5.5 Analisis de variables significativas

De esta manera, llegamos a la conclusión que las variables más significativas del modelo son:


regressor_columns

omitted_regressors = [regressor_columns[2], regressor_columns[9]]

regressor_columns_def = [regressor for regressor in regressor_columns if regressor not in omitted_regressors]

regressor_columns_def

### Variables Significantes

A continuación, hacemos un análisis de las variables que encontramos son significantes para el modelo

| **Variable** | **Significado** | **Interpretación** |
|---------------|----------------|--------------------|
| **OverallQual** | *Overall Material and Finish Quality* → Calidad general de los materiales y acabados. | Escala del **1 al 10** que evalúa la calidad general de la construcción y los acabados. Un valor alto indica **mejor calidad** y suele asociarse con precios más altos. |
| **GrLivArea** | *Ground Living Area (sq ft)* → Área habitable sobre el nivel del suelo (en pies²). | Representa el **espacio habitable total por encima del suelo**, sin incluir sótanos ni garajes. Es una de las variables más correlacionadas con `SalePrice`. |
| **GarageArea** | *Garage Area (sq ft)* → Área del garaje (en pies²). | Indica el **tamaño del garaje**, lo que refleja capacidad de estacionamiento o espacio adicional. |
| **YearBuilt** | *Original Construction Year* → Año de construcción original. | Año en que fue construida la vivienda. Las casas más nuevas tienden a tener **mayor valor**, aunque también depende del mantenimiento y remodelaciones. |
| **1stFlrSF** | *First Floor Square Feet* → Superficie del primer piso (en pies²). | Área del **primer nivel** de la vivienda. Puede correlacionarse con `GrLivArea`, pero se refiere específicamente al piso inferior. |
| **YearRemod/Add** | *Remodel Year* → Año de remodelación o ampliación. | Año en que la vivienda fue **remodelada o ampliada**. Si coincide con `YearBuilt`, significa que **no ha sido remodelada**. |
| **FullBath** | *Full Bathrooms Above Grade* → Número de baños completos sobre el nivel del suelo. | Cuenta los baños con ducha o bañera ubicados **en niveles habitables**, lo que impacta en la comodidad y el valor. |
| **GarageYrBlt** | *Year Garage was Built* → Año de construcción del garaje. | Indica cuándo se construyó el garaje. Si es nulo, la vivienda **no tiene garaje**. También refleja la antigüedad de la infraestructura. |

---





### Variables categóricas (dummies de calidad)


| **Variable** | **Significado** | **Interpretación del valor 1** |
|---------------|----------------|-------------------------------|
| **ExterQual_TA** | *Exterior Quality = Typical/Average* → Calidad exterior promedio. | La vivienda tiene un **acabado exterior de calidad típica o media**, funcional pero sin lujos. |
| **ExterQual_Gd** | *Exterior Quality = Good* → Calidad exterior buena. | La vivienda tiene un **acabado exterior de buena calidad**, mejor que el promedio y asociado a un **mayor valor de mercado**. |
| **KitchenQual_TA** | *Kitchen Quality = Typical/Average* → Calidad promedio de la cocina. | La cocina tiene materiales y acabados **de nivel medio**, ni lujosos ni deficientes. |
| **FireplaceQu_Sin categoria** | *Fireplace Quality = Sin categoría / Nulo* | Indica que la **calidad de la chimenea no fue registrada** o que la casa **no tiene chimenea**. Esta categoría fue creada durante la limpieza de datos para representar los valores faltantes. |

---

### Variables descartadas

omitted_regressors

---

| **Variable** | **Significado** | **Tipo** | **Interpretación** |
|---------------|----------------|-----------|--------------------|
| **GarageCars** | *Garage Capacity (number of cars)* → Capacidad del garaje medida en número de automóviles. | Numérica | Indica cuántos **vehículos caben en el garaje**. Es un indicador directo del tamaño y funcionalidad del garaje. Valores mayores sugieren **viviendas más amplias y de mayor valor**. Muy correlacionada con `GarageArea` y `SalePrice`. |
| **Foundation_PConc** | *Foundation = PConc (Poured Concrete)* → Tipo de **cimiento de concreto vertido** (losas o muros moldeados en sitio). | Categórica (dummy) | Toma valor **1 si la vivienda tiene cimientos de concreto vertido**, y 0 en caso contrario. Este tipo de cimiento se asocia con **mayor durabilidad, aislamiento térmico y resistencia estructural**, características de **construcciones modernas o de alta calidad**. |

---


De esta manera, calculamos nuevamente nuestro modelo, usando únicamente las variables significantes:

X = df[regressor_columns_def].to_numpy()

## inserción de 1
X_1 = np.insert(X, obj=0, values=1, axis=1)

X_1

X_1.shape

# Se recalcula el modelo con OLS
model = OLS(Y,X_1)
result = model.fit()

result.params

result.summary()

---
Tras la obtención de nuestro modelo, procedemos a guardarlo para uso futuro.

X_1.dump('../Data/X_def.dat')
result.params.dump('../Data/B_def.dat')
Y.dump('../Data/Y_def.dat')

result.save("../Data/ols_def_results.pickle")

with open("../Data/regressor_columns_def.pickle", "wb") as fp:   #Pickling
...   pickle.dump(regressor_columns_def, fp)
