==> Notebooks/09-interpretation.ipynb <==
# 9. Interpretación de Resultados

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import pickle

filename = '../Data/selected_model.sav'

loaded_model = pickle.load(open(filename, 'rb'))

filename = '../Data/scaler.sav'
scaler = pickle.load(open(filename, 'rb'))

loaded_model.coef_[0]

with open("../Data/regressor_columns_def2.pickle", "rb") as fp:   #Pickling
...   regressor_columns_def = pickle.load(fp)

df = pd.read_csv('../Data/Ames-Housing-regressor-columns.csv', delimiter=',')

regressor_columns_def

En base al modelo seleccionado, podemos hacer las siguientes conclusiones:

- El intercepto es 0. Tiene sentido que casa sin area o garajes no tenga valor, en vez de un valor negativo o positivo.
- Las variables de mayor peso son 'Overall Qual' y 'Gr Liv Area'. Tiene sentido que a mayor area vivible la casa cueste más, y a mayor calidad también se presente un incremento en el precio.
- Se tienen las siguiente variables con influencia negativa:
  1. 'Full Bath': Es contraintuivo que esta variable presente influencia negativa. A primera instancia, uno podría considerar que más baños completos deberían incrementar el precio.
  2. 'Garage Yr Blt': Puede indicar que casas más antiguas son consideradas más valiosas.
  3. 'Kitchen Qual_TA': La presencia de una cocina de calidad medio puede afectar el precio en relación con cocinas más caras y lujosas.
  4. 'Fireplace Qu_Sin categoria': Esto indica que una casa no tiene chimenea, lo cual afectar negativamente su precio
- Podrían analizarse la posibilidad de incluir más variables como estrato, número de baños, area de piscina, etc. que podría también afectar el precio.
- El uso de estandarización complica el analisis en términos prácticos de los coeficientes. Pero usando la información entregada por el scaler y recordando el hecho que algunas variables fueron transformadas con logaritmos, podemos hacer algunas conclusiones: 

Variables transformadas:

['Misc Val', 'Pool Area', 'Lot Area', 'Low Qual Fin SF', '3Ssn Porch', 'Kitchen AbvGr', 'BsmtFin SF 2', 'Enclosed Porch', 'Screen Porch', 'Bsmt Half Bath', 'Mas Vnr Area', 'Open Porch SF', 'Wood Deck SF', **'SalePrice'**, 'Lot Frontage', '1st Flr SF', 'BsmtFin SF 1', 'MS SubClass', **'Gr Liv Area'**, 'Total Bsmt SF']

df[regressor_columns_def].mean()

scaler.mean_

scaler.var_**0.5

## Ground Living area

Ground living area fue transformada logaritmicamente, por lo que cualquier cambio debe ser aplicada la misma transformación. Por ejemplo entonces, si quisieramos ver cuanto afecta un cambio en 100 ft2 en el area habitable al precio, nos daría un incremento en el valor aproximado de 4 USD.

np.expm1((np.log1p(100)/(scaler.var_[1]**0.5))*loaded_model.coef_[0][2])
